Add config option: allow-eval
This is a workaround for issue #101.

+++ index.js
@@ -100,20 +100,22 @@ manager.on("message", (shard, input) => {
 });
 
 // # Console Functions # //
-process.stdin.on("data", (text) => {
-  if(text.toString() === "stop\r\n" || text.toString() === "exit\r\n" || text.toString() === "stop\n" || text.toString() === "exit\n") {
-    manager.broadcastEval("client.destroy();")
-    .then(() => {
-      process.exit();
-    });
-  }
-  else {
-    console.log("Eval on index:");
-    try {
-      eval(text.toString());
+if(config["allow-eval"] === true) {
+  process.stdin.on("data", (text) => {
+    if(text.toString() === "stop\r\n" || text.toString() === "exit\r\n" || text.toString() === "stop\n" || text.toString() === "exit\n") {
+      manager.broadcastEval("client.destroy();")
+      .then(() => {
+        process.exit();
+      });
     }
-    catch(err) {
-      console.log(err);
+    else {
+      console.log("Eval on index:");
+      try {
+        eval(text.toString());
+      }
+      catch(err) {
+        console.log(err);
+      }
     }
-  }
-});
+  });
+}
